End of section 10

+++ Application/Activities/Delete.cs
@@ -1,3 +1,5 @@
+using Application.Core;
+using Domain;
 using MediatR;
 using Persistence;
 
@@ -5,12 +7,12 @@ namespace Application.Activities
 {
     public class Delete
     {
-        public class Command:IRequest
+        public class Command:IRequest<Result<Unit>>
         {
             public Guid Id { get; set; }
         }
 
-        public class Handler:IRequestHandler<Command>
+        public class Handler:IRequestHandler<Command, Result<Unit>>
         {
             private readonly DataContext _context;
             public Handler(DataContext context)
@@ -18,12 +20,18 @@ namespace Application.Activities
                 _context=context;
             }
 
-            public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 var activity = await _context.Activities.FindAsync(request.Id);
-                _context.Activities.Remove(activity);
-                await _context.SaveChangesAsync();
-                return Unit.Value;
+
+                if(activity==null) return null;
+
+                _context.Activities.Remove(activity) ;
+                var result = await _context.SaveChangesAsync() > 0;
+
+                if(!result) return Result<Unit>.Failure("Failed to delete the activity");
+
+                return Result<Unit>.Success(Unit.Value);
             }
         }
     }
